requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.1
pytubefix==8.12.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.47.1
triton==3.0.0
tzdata==2025.1
urllib3==2.3.0
ultralytics==8.3.66
streamlit